Major update, which changes the concept of chained tags - {{foo a {{bar b}} }} and replaces it with the syntax {{foo a bar(b)}}. You can register helper functions (example step-by-step/10) and call them from the template markup.
Added support for {{#if a===b}} and similar comparisons.

Added support for numbers, true, false, null, so you can do {{#if a===2}} -
which does number comparison and does not convert to strings. Note that all
these changes confer many of the features and scenario support availabe in
jquery.tmpl.js - but whereas there, this was achieved by eval'ing inline
expressions in the template, the JsRender does not do an eval and does not
allow inline expressions, other than with the {{* code}} tag.

Added initial support for presenters.

Fixes the following Issues
Issue 5 - in that whitespace is preserved by default
Issue 10 (comparisons)
Issue 12 (numbers)
Issue 14