v1.0.13 release
This update contains a minor bug fix:
https://github.com/BorisMoore/jsviews/issues/455: "Using a tmpl selector
that begins with a # but is not an ID selector does not work".

This change applies to jsrender and jsviews

The update also includes some minor documentation updates, and moves
to the latest versions of JQuery and JQuery UI, for all samples, examples,
and test pages.

Sync with other changes in v1.0.13 for JsViews